Birthday-Tracker-With-Alert/requirements.txt
streamlit
requests
boto3
python-dotenv

